fix(nvim): fix jump to next placeholder to use ctrl-n/p
What
---
Update the keybinding to move from one placeholder to another with
lussnip from `tab` to `ctrl-n` (next) and `ctrl-p` (previous).

Why
---
When using `tab` the editor was getting confused if it should go to the
next placeholder or just indent the code. In vim for next and previous
it's usually `ctrl-n` and `ctrl-p`.